Provide example of executeTask not executing provided task
I would expect the task output to include the word 'command' as set up
in the line:

   () => sharedState + ' command',

inside the CustomExecution inside the task created in the command
function.

However it doesn't and instead includes 'getTask' which indicates that
it has run a task provided by the 'provideTasks' method on the provider
rather than the task we have provided directly to the executeTask
function.