Provide multi-file drag to editor example
With the code as is, dragging from one of the nodes in the tree view
(eg. bb) will open tree-drag-and-drop-1.md in the editor with the
contents that we've written to the file.

The same for the second file without the first.

If we uncomment everything so that we have two files in the array then
we get an error saying that it cannot read 'tree-drag-and-drop-2.md'.