webworker: show how the scope between the extension vs. test code is not shared
In a node.js context, when running unit tests the context is shared between the
extension and unit test code. This means that the state created in the extension
context is the same as what the unit tests are using.

For example, defining a variable in the global scope of the extension code
will be accessible by the unit tests.

---

But in a web extension, when running with webworkers this context is separate
between the extension and unit tests. They only share the globalThis object from
what I could find.

---

This commit displays these differences by showing in the unit tests that there
is a disconnect in globally defined objects in a web context.

Signed-off-by: nkomonen <nkomonen@amazon.com>